Make BslLeafletVideo components
